proyecto_vehiculos_django/requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
crispy-bootstrap5==0.6
Django==4.0.5
django-bootstrap-v5==1.0.11
django-colorfield==0.9.0
django-crispy-forms==1.14.0
Pillow==10.0.0
python-slugify==8.0.1
signals==0.0.2
soupsieve==2.4.1
sqlparse==0.4.4
text-unidecode==1.3
tzdata==2023.3
